ci: consolidate playwright config to one file
