+++ package.json
@@ -14,8 +14,8 @@
     "prettier": "prettier --check '**/*'",
     "stylelint": "stylelint '**/*.scss'",
     "test": "jest",
-    "playwright": "npx playwright test --config=playwright.dev.config.ts",
-    "playwright:debug": "PWDEBUG=1 npx playwright test --config=playwright.dev.config.ts",
+    "playwright": "IS_DEV=true npx playwright test",
+    "playwright:debug": "IS_DEV=true PWDEBUG=1 npx playwright test",
     "jest:debug": "node --inspect-brk node_modules/.bin/jest --runInBand --no-cache",
     "compile": "tsc",
     "lint": "yarn eslint --max-warnings=0 && yarn prettier && yarn stylelint",
