+++ playwright.config.ts
@@ -1,13 +1,25 @@
 // playwright.config.ts
 import { type PlaywrightTestConfig, devices } from "@playwright/test"
 
-const BASE_URL = process.env.URL
+const { IS_DEV } = process.env
+const LOCALHOST = "http://localhost:6006/"
+const BASE_URL = IS_DEV ? LOCALHOST : `https://${process.env.URL}`
+
+const devConfig: PlaywrightTestConfig = {
+  webServer: {
+    command: "yarn storybook",
+    url: LOCALHOST,
+    timeout: 120 * 1000000,
+    reuseExistingServer: true,
+  },
+}
 
 const config: PlaywrightTestConfig = {
   testMatch: ["**/*.playwright.spec.ts?(x)"],
-  retries: 1,
+  retries: IS_DEV ? 0 : 1,
   use: {
-    baseURL: `https://${BASE_URL}`,
+    baseURL: BASE_URL,
+    trace: IS_DEV ? "on-first-retry" : "off",
   },
   projects: [
     {
@@ -16,6 +28,7 @@ const config: PlaywrightTestConfig = {
     },
   ],
   outputDir: "./playwright/test-results",
+  ...(IS_DEV ? devConfig : {}),
 }
 
 export default config
